conv_num prefix suffix pram

+++ ft_intlen_base.c
@@ -6,13 +6,13 @@
 /*   By: mhaziza <mhaziza@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/23 16:42:34 by mhaziza           #+#    #+#             */
-/*   Updated: 2016/11/23 21:48:32 by mhaziza          ###   ########.fr       */
+/*   Updated: 2017/01/14 16:22:44 by mhaziza          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_intlen_base(unsigned int n, int base)
+size_t	ft_intlen_base(unsigned long n, int base)
 {
 	size_t			i;
